+++ src/main/java/com/twilio/rest/api/v2010/account/sip/credentiallist/CredentialUpdater.java
@@ -51,9 +51,9 @@ public class CredentialUpdater extends Updater<Credential> {
     }
 
     /**
-     * The password that the username will use when when authenticating SIP
-     * requests. The password must be a minimum of 12 characters, contain at least 1
-     * digit, and have mixed case. (eg “IWasAtSignal2018”).
+     * The password that the username will use when authenticating SIP requests. The
+     * password must be a minimum of 12 characters, contain at least 1 digit, and
+     * have mixed case. (eg “IWasAtSignal2018”).
      * 
      * @param password The password will not be returned in the response.
      * @return this